% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_likert_data.R
\name{dummy_likert_data}
\alias{dummy_likert_data}
\title{Dummy Daten}
\usage{
dummy_likert_data(
  n = 100,
  likert_levels = c("Strongly disagree", "Disagree", "Neither agree nor disagree",
    "Agree", "Strongly agree")
)
}
\arguments{
\item{n}{integer. Anzahl}

\item{likert_levels}{character.}
}
\value{
data.frame
}
\description{
Dummy Daten
}
