% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-likertplot.R
\name{gg_likertplot}
\alias{gg_likertplot}
\alias{gg_likert_stacked}
\title{ggplot-Barcharts for Likert and Multi-Response}
\usage{
gg_likertplot(
  data,
  ...,
  facet_formula = NULL,
  include.order = NULL,
  decreasing = FALSE,
  include.reference = NULL,
  include.percent = TRUE,
  exclude_fill_values = NULL,
  reverse_likert = FALSE,
  facet_label_wrap = 50,
  y_label_wrap = 50,
  palette = "RdBu",
  direction = 1,
  width = 0.9
)

gg_likert_stacked(
  data,
  ...,
  facet_formula = NULL,
  include.order = FALSE,
  decreasing = FALSE,
  palette = "BrBG",
  direction = 1,
  facet_label_wrap = 50,
  include.value = TRUE,
  include.percent = TRUE,
  labels_size = 3.5,
  labels_color = "black",
  labels_hide_below = 0.05,
  y_label_wrap = 50,
  width = 0.9
)
}
\arguments{
\item{data}{data.frame im long Format}

\item{...}{to Summarise_likert()}

\item{facet_formula}{formula. ggplot NULL or formula Item ~ .}

\item{include.order, decreasing}{character. include.order = c("right", "left")  die Sortierung}

\item{include.reference}{numeric. Referenzline}

\item{include.percent}{logical.}

\item{exclude_fill_values}{character. ggstats ausgeschlossene Level}

\item{reverse_likert}{logical.  ggstats nicht zu Verwenden Balken umdrehen gethaber nicht für die Beschriftung}

\item{y_label_wrap, facet_label_wrap}{ggstats numeric. Umbrechen der Labels}

\item{palette, direction}{scharacter. ggstats  scale_fill_brewer  direction = 1 or -1}

\item{width}{numeric.  ggstats breite der Balken}

\item{include.value}{logical. Prozent anzeigen}

\item{labels_size}{numeric. Schriftgrösse}

\item{labels_color}{character.   Zweifarbig  labels_color = c("blue","gray")}

\item{labels_hide_below}{numeric. minimum der Beschriftung}
}
\value{
ggplot
}
\description{
ggplot-Barcharts for Likert and Multi-Response
}
\examples{
 
Likert <- dummy_likert_data()
 
Likert |>
  Summarise_likert_long(q1, q2, q3, q4, q5, q6, q7, q8, q9,
                        by =~ Sex,
                        grouping = list(
                          FC.2 = c("q1", "q2"),
                          FC.3 = c("q3","q4", "q5", "q6"),
                          FC.4 = c("q7", "q8", "q9")
                        )
  ) |>
  gg_likertplot()


Multi <- stp25likert::dummy_multi_data()

Multi |>
  stp25likert::Tbll_multi(q1, q2, q3, q4, q5, q6, q7, by = ~ Sex) |>
  gg_likert_stacked(
    include.order = "l",
    palette = "BrBG", direction = 1,
    
    labels_size = 4.5,
    labels_color = c("blue","gray"),
    width = .5)

}
