% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_data.R
\name{likert_data}
\alias{likert_data}
\alias{multi_data}
\title{likert data}
\usage{
likert_data(
  x,
  ...,
  grouping = NULL,
  include.order = TRUE,
  reverse.levels = FALSE,
  decreasing = TRUE,
  use.level = NULL
)

multi_data(..., use.level = 1, reverse.levels = FALSE)
}
\arguments{
\item{x}{Tbll_likert -Objekt}

\item{...}{platzhalter}

\item{reverse.levels}{= FALSE, levels umdrehen}

\item{include.total}{margin an Summarise}
}
\value{
long tibble mit Item, levels,  Freq
}
\description{
You could use `gglikert_data()` to just produce the dataset to be plotted.
}
\examples{

#' require(stp25likert)
set.seed(1)
n <- 100
lvs <- c("--", "-", "o", "+", "++")
DF2 <- data.frame(
  Magazines = cut(rnorm(n), 5, lvs),
  Comic.books = cut(rnorm(n), 5, lvs),
  Fiction = cut(rnorm(n), 5, lvs),
  Newspapers = cut(rnorm(n), 5, lvs),
  Geschlecht = cut(rnorm(n), 2, c("m", "f"))
) |>  stp25tools::Label(Comic.books = "Comic-books")

dat_lg <- Tbll_likert(
  DF2,
  Magazines,
  Comic.books,
  Fiction,
  Newspapers,
  by =  ~ Geschlecht,
  include.order = TRUE
)

likert_data(dat_lg)

likert_data(
  DF2,
  Magazines,
  Comic.books,
  Fiction,
  Newspapers,
  by =  ~ Geschlecht,
  include.total =TRUE
)

lvls <- c(TRUE, FALSE)
n<- 150
DF_multi <-
  tibble(
    q1 =  rbinom(n, 1, .5) ,
    q2 =   rbinom(n, 1, .4),
    q3 =   rbinom(n, 1, .55),
    q4 =   rbinom(n, 1, .56),
    q5 =   rbinom(n, 1, .53),
    q6 =   rbinom(n, 1, .2),
    q7 =   rbinom(n, 1, .85),
    q8 =  rbinom(n, 1, .95),
    q9 =  rbinom(n, 1, .25)
  ) |>
  mutate(across(everything(), ~ factor(.x ))) |>
  mutate(Sex = sample(c("male", "female"), n, replace = TRUE),
         Age = sample(c("18-30", "30-50", ">50"), n, replace = TRUE)) |>
  stp25tools::Label(
    # FC.2
    q1 =  "Nuts.Seeds",
    q2 =  "Legumes",
    # FC.3
    q3 = "Vegetables Juice ",
    q4 = "Fruit",
    q5 = "Vegetables",
    #   FC.4
    q6 = "Milk",
    q7 = "Cheese",
    q8 = "Yoghurt",
    q9 = "Dairy Alternatives"
  )

summary(DF_multi)

DF_multi |>
  likert_data(q1,q2, q3, q4, q5, q6, q7, q8, q9,
              by=~ Sex + Age,
              grouping= list(
                FC.2 = c("q1", "q2"),
                FC.3 = c("q4", "q5", "q6"),
                FC.4 = c("q7", "q8", "q9")
              )
  )

}
