% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertplot.R
\name{likertplot}
\alias{likertplot}
\alias{likert_plot}
\title{Barcharts for Likert}
\usage{
likertplot(
  x = Item ~ .,
  data = NULL,
  main = "",
  ylab = "",
  sub = "",
  xlab = if (horizontal) {
     if (as.percent) 
         "Prozent"
     else "Anzahl"
 }
    else "",
  ylim = NULL,
  xlim = NULL,
  col = NULL,
  rightAxis = FALSE,
  positive.order = NULL,
  include.order = NULL,
  decreasing = TRUE,
  as.percent = TRUE,
  auto.key = list(space = space, columns = columns, between = 1),
  ReferenceZero = NULL,
  reference.line.col = "gray65",
  col.strip.background = "gray97",
  wrap = TRUE,
  columns = 2,
  space = "top",
  horizontal = TRUE,
  between = list(x = 1 + (horizontal), y = 0.5 + 2 * (!horizontal)),
  par.settings = NULL,
  ...
)

likert_plot(
  ...,
  main = "",
  ylab = "",
  sub = "",
  xlab = if (as.percent) "Prozent" else "Anzahl",
  ylim = NULL,
  xlim = NULL,
  type = 1,
  col = NULL,
  rightAxis = FALSE,
  as.percent = TRUE,
  auto.key = list(space = space, columns = columns, between = 1),
  ReferenceZero = include.reference,
  reference.line.col = "gray65",
  col.strip.background = "gray97",
  wrap = TRUE,
  columns = 2,
  space = "top",
  horizontal = TRUE,
  positive.order = NULL,
  between = list(x = 1 + (horizontal), y = 0.5 + 2 * (!horizontal)),
  par.strip.text = list(lines = 1, cex = 0.8),
  par.settings = NULL,
  include.reference = NULL,
  include.total = FALSE,
  relevel = NULL,
  relabel = NULL,
  include.order = NULL,
  decreasing = TRUE,
  caption = "",
  include.table = FALSE,
  include.mean = TRUE,
  include.n = FALSE,
  include.na = FALSE,
  include.percent = TRUE,
  include.count = TRUE
)
}
\arguments{
\item{x}{formula}

\item{data}{daten
Format tibble (wide)
sex Item Strongly Neither Agree}

\item{main, ylab, sub, xlab}{Beschriftung}

\item{col}{HH::brewer.pal.likert}

\item{rightAxis, as.percent, ReferenceZero, reference.line.col, col.strip.background}{an HH:::plot.likert.formula}

\item{positive.order}{das nicht verwenden!! - wird ueber Tbll_likert oder include.order = TRUE gesteuert.}

\item{include.order, decreasing}{Sortieren der Items}

\item{auto.key, columns, space}{columns = 2,}

\item{wrap}{Zeien Umbrechen farbe("likert.blue.red", data$nlevels, middle = ReferenceZero)}

\item{...}{an stp25stat2::Likert}

\item{type}{nur in likert_plot Bei Gruppen Items als Zeilen => 1, oder Gruppen als Zeilen => 2}

\item{par.strip.text}{an  HH:::plot.likert.formula}

\item{include.total}{an stp25stat2:::Likert}

\item{relevel}{Uberschreibt die levels levels(x) <- relevel ist nur nur in likert_plot vorhanden}

\item{relabel}{ändert die Sortierung}

\item{include.table, include.mean, include.n, include.percent, include.count, include.na, caption}{Tabelle ausgeben (output ist eine Altlast)}

\item{include.order}{sortiere muss die länge der Items entsprechen}
}
\value{
lattice Plot

HH likertplot (lattice-Plot)
}
\description{
copie from HH:::plot.likert.formula
Constructs and plots diverging stacked barcharts for Likert

likert_plot: indirekt likertplot()  mit hilfe von prepare_data2
}
\details{
Die orginale Funktion hat bei der Sortierung (positive.order) einen Fehler.
}
\examples{
#' 
DF <- dummy_likert_data()

rslt <- DF  |>
  Tbll_likert(
    q1,q2,q3,q4,q5,q6,
    by = ~ sex,
    include.order = TRUE
  )

# attr(rslt, "plot")$results

likertplot(Item ~ . | sex,
           data = rslt,
           between=list(x=0))

# col = likert_col(attr(data, "plot")$nlevels, middle = ReferenceZero)

DF  |> likert_plot(q1,q2,q3,
                   relevel = letters[1:5],
                   ReferenceZero = 1.5,
                   columns=5)

  DF2<- DF |> 
Summarise_likert_wide(q1, q2, q3, q4, q5, q6, q7, q8, q9,
                      # by =~ sex,
                      grouping = list(
                        FC.2 = c("q1", "q2", "q6"),
                        FC.3 = c("q3","q4", "q5"),
                        FC.4 = c("q7", "q8", "q9")
                      )
)  

# require(lattice)
likertplot(Item ~ . | .grouping , DF2,
           scales=list(y=list(relation="free")),
           layout=c(1,3),
           #  positive.order=TRUE,
           between=list(y=0),
           strip=FALSE,
           strip.left=strip.custom(bg="gray97"),
           par.strip.text=list(cex=.8, lines=5),
           main="Is your job professionally challenging?",
           ylab=NULL,
           wrap =FALSE )
           
}
