% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_weighted_mean.R
\name{order_weighted_mean}
\alias{order_weighted_mean}
\title{Sortieren von Likert-Data}
\usage{
order_weighted_mean(
  data,
  x,
  include.reference = NULL,
  m_weight = NULL,
  measure = setdiff(names(data), all.vars(x)),
  items = intersect(all.vars(x[[2]]), names(data)),
  groups = intersect(all.vars(x[[3]]), names(data)),
  nlevels = ncol(data) - length(all.vars(x)) + 1,
  ReferenceZero = include.reference,
  include.order = FALSE,
  decreasing = TRUE,
  dat_attr = attr(data, "tbll")
)
}
\arguments{
\item{data}{.data.frame Summarise_likert -Objekt tibble mit attribut "tbll"}

\item{x}{formula. wen nicht mit Summarise_likert erstellt dann die Formel}

\item{include.reference, ReferenceZero}{numeric.neutraler Wert}

\item{m_weight}{numeric. kann verändert werden internal}

\item{measure}{null. or character  internal}

\item{items}{null. or character  internal}

\item{groups}{null. or character internal}

\item{nlevels}{null. or character internal}

\item{include.order}{.logical or character richtung der sortierung}

\item{decreasing}{logical. zu or abnahme}

\item{dat_attr}{list. internal}
}
\value{
tibble
}
\description{
Sortieren der Items nach Mittelwert. Items werden als factor() geordnet.
}
\examples{

Likert <- dummy_likert_data()

Likert |>
  Summarise_likert_long(q1,q2,q3,q4,q5,q6,q7,q8,q9,
                        grouping = list(
                          FC.2 = c("q1", "q9", "q2", "q5"),
                          FC.3 = c("q3", "q4", "q6"),
                          FC.4 = c("q7", "q8")
                        )) |> 
 order_weighted_mean(include.reference=4, include.order = "l")
 
}
