% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{stp25likert-package}
\alias{stp25likert}
\alias{stp25likert-package}
\title{stp25likert: Analysis of Likert-type items.}
\description{
The purpose of stp25stat2 is to facilitate work and provide functions for creating statistical evaluation tables.
}
\author{
\strong{Maintainer}: Wolfgang Peter \email{w.peter@statistik-peter.at}

}
\keyword{internal}
