% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_data.R
\name{Summarise_likert}
\alias{Summarise_likert}
\alias{Summarise_likert_wide}
\alias{Summarise_multi}
\title{Summary likert data}
\usage{
Summarise_likert(
  x,
  ...,
  grouping = NULL,
  include.order = TRUE,
  reverse.levels = FALSE,
  decreasing = TRUE,
  use.level = NULL
)

Summarise_likert_wide(...)

Summarise_multi(..., use.level = 1, reverse.levels = FALSE)
}
\arguments{
\item{x}{Tbll_likert -Objekt}

\item{...}{platzhalter}

\item{grouping}{list.}

\item{include.order}{logical.}

\item{reverse.levels}{logical.  levels umdrehen}

\item{decreasing}{logical.}

\item{use.level}{logical.}
}
\value{
long tibble mit Item, levels,  Freq
}
\description{
Summe der Items als Long-Datafile.
}
\details{
You could use `gglikert_data()` to just produce the dataset to be plotted.
}
\examples{
 
DF <- dummy_likert_data()

rslt <- DF  |> 
  Tbll_likert(
    q1,q2,q3,q4,q5,q6,
    by = ~ sex,
    include.order = TRUE
  )
Summarise_likert(rslt)

DF |> 
  Summarise_likert(q1, q2, q3, q4, q5, q6, q7, q8, q9,
              by =~ sex + age,
              grouping = list(
                FC.2 = c("q1", "q2"),
                FC.3 = c("q4", "q5", "q6"),
                FC.4 = c("q7", "q8", "q9")
              )
  )
}
