% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_likert.R
\name{Tbll_likert}
\alias{Tbll_likert}
\alias{Tbll_likert.data.frame}
\alias{Tbll_likert.likert}
\alias{Likert}
\alias{print.likert}
\title{Tbll_likert}
\usage{
Tbll_likert(...)

\method{Tbll_likert}{data.frame}(
  x,
  ...,
  include.reference = NULL,
  include.mean = TRUE,
  include.n = FALSE,
  include.na = FALSE,
  include.order = FALSE,
  include.percent = TRUE,
  include.count = TRUE,
  include.total = FALSE,
  exclude.levels = NULL,
  decreasing = TRUE,
  labels = c("low", "neutral", "high"),
  reverse.levels = FALSE,
  reorder.levels = NA
)

\method{Tbll_likert}{likert}(
  x,
  include.reference = NULL,
  include.mean = TRUE,
  include.n = FALSE,
  include.na = FALSE,
  include.order = FALSE,
  include.percent = TRUE,
  include.count = TRUE,
  include.total = FALSE,
  labels = c("low", "neutral", "high"),
  decreasing = TRUE,
  ...
)

Likert(
  ...,
  labels = NULL,
  reverse.levels = FALSE,
  reorder.levels = NA,
  include.total = FALSE,
  exclude.levels = NULL
)

\method{print}{likert}(x, ...)
}
\arguments{
\item{...}{alles an prepare.data}

\item{x}{data.frame or Likert_objekt}

\item{include.reference, labels}{numeric include.reference = 2 (drei Gruppen)
include.reference = 2.5 (zwei Gruppen)
 Neutrales Element in Kombination mit
 labels = c("low", "neutral", "high")}

\item{include.mean, include.n}{Zusatz Ergebnisse}

\item{include.na}{logical. Missing: hier werden die Prozent inclusive der NAs berechnet.}

\item{include.order, decreasing}{sortierung nach mittelwert}

\item{include.percent, include.count}{Format Prozent/Anzahl}

\item{include.total}{An Summarise_likert und Likert: logical oder string  zB. include.total ="Alle"}

\item{exclude.levels}{An Summarise_likert und position des zu excludierenden levels 
(zb 'exclude.levels = 5' ist das gleiche wie 'reorder.levels = -5')}

\item{reverse.levels}{An Summarise_likert und Likert: logical  rev(item)}

\item{reorder.levels}{An Summarise_likert und Likert: integer factor(item, levels(item)[reorder.levels])}
}
\value{
data.frame mit attributen fuer plotlikert

liste mit  results = data, sowie m, sd, n
}
\description{
Analysis of Likert-type items.

Likert: Auszählen der Häufigkeiten Obsolet neu ist Summarise_likert()
}
\examples{
#
Likert <- dummy_likert_data(245)

Likert |>
  Tbll_likert(q1, q2, q3, q4, by = ~ Sex, include.order = TRUE)

Likert |>
  Tbll_likert(q1, q2, q3, q4, include.order = TRUE)

Likert |>
  Tbll_likert(q1,
              q2,
              q3,
              q4,
              include.percent = FALSE,
              include.mean = FALSE) |>
#  stp25output2::Output() |>
  likertplot()


}
