% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbll_multi.R
\name{Tbll_multi}
\alias{Tbll_multi}
\title{Analysis multi-response items.}
\usage{
Tbll_multi(
  ...,
  by = NULL,
  digits = 0,
  include.order = TRUE,
  exclude.last.order = FALSE,
  use.level = 1,
  include.label = TRUE,
  include.n = TRUE,
  include.nr = FALSE,
  include.total = FALSE,
  include.test = FALSE,
  include.normality.tests = FALSE,
  include.custom = NULL,
  include.value = NULL
)
}
\arguments{
\item{...}{an prepate.data stp25stat2::Tbll_desc_multi}

\item{by}{an prepate.data}

\item{digits}{numeric.}

\item{use.level, include.label}{.logical Beschriftung}

\item{include.n, include.nr}{logical Anzahl}

\item{include.total, include.order, exclude.last.order}{logical.}

\item{include.test, include.normality.tests, include.custom, include.value}{logical. stp25stat2::Tbll_desc_multi}
}
\description{
Tbll_multi() ist praktisch die Funktio stp25stat2::Tbll_desc() mit Atribut attr(rslt, "tbll_likert").
}
\examples{

Multi <- dummy_multi_data()

Multi |>
  Tbll_multi(q1, q2, q3, q4, by = ~ Sex) |>
 # stp25output2::Output() |> 
  likert_stacked()

}
