% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_data.R
\name{dummy_likert_data}
\alias{dummy_likert_data}
\alias{dummy_multi_data}
\title{Dummy Daten}
\usage{
dummy_likert_data(
  n = 2 * 3 * 9 * 3,
  levels = c("Strongly disagree", "Disagree", "Neither agree nor disagree", "Agree",
    "Strongly agree")
)

dummy_multi_data(n = 100)
}
\arguments{
\item{n}{integer. Anzahl}

\item{levels}{character.}
}
\value{
data.frame
}
\description{
Dummy Daten
}
