% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice-likertplot.R
\name{likertplot}
\alias{likertplot}
\alias{likertplot.data.frame}
\alias{likertplot.formula}
\alias{likertplot.default}
\title{Barcharts for Likert}
\usage{
likertplot(...)

\method{likertplot}{data.frame}(
  data,
  formula = NULL,
  include.reference = NULL,
  include.total = FALSE,
  use.level = NULL,
  reverse.levels = FALSE,
  nlevels = NULL,
  m_weight = NULL,
  groups = NULL,
  ReferenceZero = include.reference,
  ...
)

\method{likertplot}{formula}(formula, data = NULL, ...)

\method{likertplot}{default}(
  x,
  data,
  items,
  measure,
  groups,
  grouping = FALSE,
  nlevels,
  ReferenceZero = NULL,
  include.order = NULL,
  decreasing = TRUE,
  main = "",
  ylab = "",
  sub = "",
  xlab = if (horizontal) {
     if (as.percent) 
         "Prozent"
     else "Anzahl"
 }
    else "",
  ylim = NULL,
  xlim = NULL,
  col = NULL,
  rightAxis = FALSE,
  as.percent = TRUE,
  columns = 2,
  space = "top",
  between = list(x = 1 + (horizontal), y = 0.5 + 2 * (!horizontal)),
  auto.key = list(space = space, columns = columns, between = 1),
  reference.line.col = "gray65",
  col.strip.background = "gray97",
  horizontal = TRUE,
  par.settings = NULL,
  wrap = TRUE,
  positive.order = NULL,
  ...
)
}
\arguments{
\item{...}{HH:::plot.likert.formula  between=list(x=0))}

\item{data}{daten
Format tibble (wide)
sex Item Strongly Neither Agree}

\item{x}{formula}

\item{include.order, decreasing}{logical or character, Sortieren der Items}

\item{main, ylab, sub, xlab}{Beschriftung}

\item{col}{HH::brewer.pal.likert}

\item{rightAxis, as.percent, ReferenceZero, reference.line.col, col.strip.background}{an HH:::plot.likert.formula}

\item{auto.key, columns, space}{columns = 2,}

\item{wrap}{Zeien Umbrechen farbe("likert.blue.red", data$nlevels, middle = ReferenceZero)}

\item{positive.order}{das nicht verwenden!! - wird ueber Tbll_likert oder include.order = TRUE gesteuert.}
}
\value{
lattice Plot
}
\description{
copie from HH:::plot.likert.formula
Constructs and plots diverging stacked barcharts for Likert
}
\details{
Die orginale Funktion hat bei der Sortierung (positive.order) einen Fehler.
}
\examples{
  
Likert <- dummy_likert_data(245)

Likert |>
  Summarise_likert(q1,q2,q3,q4,q5,q6,q7,q8,q9,
                   grouping = list(
                     FC.2 = c("q1", "q9", "q2", "q5"),
                     FC.3 = c("q3", "q4", "q6"),
                     FC.4 = c("q7", "q8")
                   )) |>
  transform(Item= as.character(Item) )|>
  plyr::arrange(.grouping)  |>
  likertplot(
    Item ~ . | .grouping,
    scales = list(y = list(relation = "free")),
    layout = c(1, 3),
    between = list(y = 0),
    strip = FALSE,
    strip.left = lattice::strip.custom(bg = "gray97"),
    par.strip.text = list(cex = .8, lines = 5),
    main = "Ernärung",
    ylab = NULL,
    wrap = FALSE
  )

Likert |>
  Summarise_likert_long(q1,q2,q3,q4,q5,q6,q7,q8,q9,
                        grouping = list(
                          FC.2 = c("q1", "q9", "q2", "q5"),
                          FC.3 = c("q3", "q4", "q6"),
                          FC.4 = c("q7", "q8")
                        )) |>
  transform(Item = as.character(Item)) |>
  plyr::arrange(.grouping)  -> dat
dat
HH::likert(
  Item ~ levels | .grouping ,
  value = "Freq",
  data = dat,
  ylab = NULL,
  scales = list(y = list(relation = "free")),
  layout = c(1, 3)
)



# p <-
# Likert |>
#   Summarise_likert(q1,q2,q3,q4,q5,q6,q7,q8,q9,
#                  by =~ Sex,
#                  grouping = list(
#                    FC.2 = c("q1", "q9", "q2", "q5"),
#                    FC.3 = c("q3", "q4", "q6"),
#                    FC.4 = c("q7", "q8")
#                  )) |>
#   transform(Item= as.character(Item) )|>
#   plyr::arrange(.grouping)  |>
#   stp25likert::likertplot(
#     Item ~ . |  Sex + .grouping ,
#     scales = list(y = list(relation = "free")),
#     layout = c(1, 3),
#     between = list(y = 0),
#     #   strip = FALSE,
#     #  strip.top = strip.custom(bg = "gray97"),
#     #  strip.left = strip.custom(bg = "gray97"),
#     #  par.strip.text = list(cex = .8, lines = 5),
#     main = "Ernärung",
#     ylab = NULL,
#     wrap = FALSE
#   )
# 
# require(latticeExtra)
# useOuterStrips(
#  p,
#  strip.left = strip.custom(horizontal = FALSE),
#  strip.lines = 2,
#  strip.left.lines = 1
# )
}
