% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{stp25likert-package}
\alias{stp25likert}
\alias{stp25likert-package}
\title{stp25likert: Analysis of Likert-type and multi-response items.}
\description{
The purpose of stp25likert is to make work easier by providing functions for creating Likert tables and plots.
}
\author{
\strong{Maintainer}: Wolfgang Peter \email{w.peter@statistik-peter.at}

}
\keyword{internal}
